requirements.txt
flask
google-generativeai
ollama
